# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agent-Sentinel Immutable Audit Ledger API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

/**
 * Raw event data (JSONB)
 */
export type AuditLogPayload = { [key: string]: unknown };

export interface AuditLog {
  /** UUID primary key */
  id: string;
  /** UTC timestamp of the event */
  timestamp: string;
  /** Identifier of the AI agent */
  agentId: string;
  /** Groups related events into a single decision trace */
  traceId: string;
  /** One of: Intent, Action, Result, Error */
  eventType: string;
  /** Raw event data (JSONB) */
  payload: AuditLogPayload;
  /**
   * Chain-of-thought rationale for the action
   * @nullable
   */
  rationale?: string | null;
  /** SHA-256 hash of this entry */
  currentHash: string;
  /**
   * Hash of the previous entry in chain
   * @nullable
   */
  previousHash?: string | null;
  /** Whether this entry was flagged as anomalous */
  isAnomalous: boolean;
  /**
   * Reason this entry was flagged as anomalous
   * @nullable
   */
  anomalyReason?: string | null;
}

export type CreateAuditLogBodyPayload = { [key: string]: unknown };

export interface CreateAuditLogBody {
  agentId: string;
  traceId: string;
  /** One of: Intent, Action, Result, Error */
  eventType: string;
  payload: CreateAuditLogBodyPayload;
  /** @nullable */
  rationale?: string | null;
}

export interface AuditLogList {
  logs: AuditLog[];
  total: number;
  limit: number;
  offset: number;
}

export interface TraceView {
  traceId: string;
  agentId: string;
  startTime: string;
  /** @nullable */
  endTime?: string | null;
  /** derived: success, error, in-progress */
  status: string;
  events: AuditLog[];
  totalEvents: number;
  hasAnomalies: boolean;
}

export type DashboardStatsEventTypeCounts = { [key: string]: number };

export interface DashboardStats {
  totalLogs: number;
  totalAgents: number;
  totalTraces: number;
  anomalyCount: number;
  eventTypeCounts: DashboardStatsEventTypeCounts;
  recentActivity: AuditLog[];
  integrityOk: boolean;
}

export type AgentListAgentsItem = {
  agentId: string;
  totalEvents: number;
  lastSeen: string;
  anomalyCount: number;
};

export interface AgentList {
  agents: AgentListAgentsItem[];
}

export interface ComplianceReport {
  agentId: string;
  startTime: string;
  endTime: string;
  generatedAt: string;
  totalEvents: number;
  anomalyCount: number;
  integrityVerified: boolean;
  logs: AuditLog[];
}

export interface IntegrityStatus {
  ok: boolean;
  /** @nullable */
  lastVerifiedAt?: string | null;
  totalChecked: number;
  tamperDetected: boolean;
  /** IDs of entries where hash chain is broken */
  tamperedEntries: string[];
  message: string;
}

export type GetAuditLogsParams = {
  /**
   * Filter by agent ID
   * @nullable
   */
  agentId?: string | null;
  /**
   * Filter by trace ID
   * @nullable
   */
  traceId?: string | null;
  /**
   * Filter by event type (Intent/Action/Result/Error)
   * @nullable
   */
  eventType?: string | null;
  /**
   * ISO 8601 start timestamp filter
   * @nullable
   */
  startTime?: string | null;
  /**
   * ISO 8601 end timestamp filter
   * @nullable
   */
  endTime?: string | null;
  /**
   * Number of results to return (default 50, max 200)
   * @nullable
   */
  limit?: number | null;
  /**
   * Pagination offset
   * @nullable
   */
  offset?: number | null;
  /**
   * If true, only return flagged anomalous entries
   * @nullable
   */
  anomaliesOnly?: boolean | null;
};

export type ExportComplianceReportParams = {
  agentId: string;
  /**
   * ISO 8601 start timestamp
   */
  startTime: string;
  /**
   * ISO 8601 end timestamp
   */
  endTime: string;
  /**
   * Export format - csv or json (default json)
   * @nullable
   */
  format?: string | null;
};