# lib/api-zod/src/generated/types/agentList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agent-Sentinel Immutable Audit Ledger API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentListAgentsItem } from "./agentListAgentsItem";

export interface AgentList {
  agents: AgentListAgentsItem[];
}