# lib/api-zod/src/generated/types/auditLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agent-Sentinel Immutable Audit Ledger API
 * OpenAPI spec version: 0.1.0
 */
import type { AuditLogPayload } from "./auditLogPayload";

export interface AuditLog {
  /** UUID primary key */
  id: string;
  /** UTC timestamp of the event */
  timestamp: string;
  /** Identifier of the AI agent */
  agentId: string;
  /** Groups related events into a single decision trace */
  traceId: string;
  /** One of: Intent, Action, Result, Error */
  eventType: string;
  /** Raw event data (JSONB) */
  payload: AuditLogPayload;
  /**
   * Chain-of-thought rationale for the action
   * @nullable
   */
  rationale?: string | null;
  /** SHA-256 hash of this entry */
  currentHash: string;
  /**
   * Hash of the previous entry in chain
   * @nullable
   */
  previousHash?: string | null;
  /** Whether this entry was flagged as anomalous */
  isAnomalous: boolean;
  /**
   * Reason this entry was flagged as anomalous
   * @nullable
   */
  anomalyReason?: string | null;
}