# lib/api-zod/src/generated/types/auditLogList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agent-Sentinel Immutable Audit Ledger API
 * OpenAPI spec version: 0.1.0
 */
import type { AuditLog } from "./auditLog";

export interface AuditLogList {
  logs: AuditLog[];
  total: number;
  limit: number;
  offset: number;
}