# lib/api-zod/src/generated/types/complianceReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agent-Sentinel Immutable Audit Ledger API
 * OpenAPI spec version: 0.1.0
 */
import type { AuditLog } from "./auditLog";

export interface ComplianceReport {
  agentId: string;
  startTime: string;
  endTime: string;
  generatedAt: string;
  totalEvents: number;
  anomalyCount: number;
  integrityVerified: boolean;
  logs: AuditLog[];
}