# lib/api-zod/src/generated/types/createAuditLogBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agent-Sentinel Immutable Audit Ledger API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAuditLogBodyPayload } from "./createAuditLogBodyPayload";

export interface CreateAuditLogBody {
  agentId: string;
  traceId: string;
  /** One of: Intent, Action, Result, Error */
  eventType: string;
  payload: CreateAuditLogBodyPayload;
  /** @nullable */
  rationale?: string | null;
}