# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agent-Sentinel Immutable Audit Ledger API
 * OpenAPI spec version: 0.1.0
 */
import type { AuditLog } from "./auditLog";
import type { DashboardStatsEventTypeCounts } from "./dashboardStatsEventTypeCounts";

export interface DashboardStats {
  totalLogs: number;
  totalAgents: number;
  totalTraces: number;
  anomalyCount: number;
  eventTypeCounts: DashboardStatsEventTypeCounts;
  recentActivity: AuditLog[];
  integrityOk: boolean;
}