# lib/api-zod/src/generated/types/integrityStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agent-Sentinel Immutable Audit Ledger API
 * OpenAPI spec version: 0.1.0
 */

export interface IntegrityStatus {
  ok: boolean;
  /** @nullable */
  lastVerifiedAt?: string | null;
  totalChecked: number;
  tamperDetected: boolean;
  /** IDs of entries where hash chain is broken */
  tamperedEntries: string[];
  message: string;
}