# lib/api-zod/src/generated/types/traceView.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agent-Sentinel Immutable Audit Ledger API
 * OpenAPI spec version: 0.1.0
 */
import type { AuditLog } from "./auditLog";

export interface TraceView {
  traceId: string;
  agentId: string;
  startTime: string;
  /** @nullable */
  endTime?: string | null;
  /** derived: success, error, in-progress */
  status: string;
  events: AuditLog[];
  totalEvents: number;
  hasAnomalies: boolean;
}